requirements.txt
pandas
numpy
plotly
ipywidgets
yfinance
python-dateutil
PyYAML
requests
